requirements.txt
farm-haystack[all]==1.26.4
flask==3.1.0
faiss-cpu==1.7.2
gunicorn==23.0.0
transformers==4.49.0